Overhaul Animations & color scheme for Music Section
Made new neon and flicker animations for both the professional and music websites.
You can see the similarities between the 2, and the themes correspond to the website's purpose.

paint splatters on all white for dj, and sleek white letters on all black for professional.

Always worth noting that the dj website links to my professional, but the professional will never link back or to the dj site. with no navbar, this keeps the guest contained.